# Improves performance of Local MD potentials