--- a/timemachine/cpp/src/kernels/k_flat_bottom_bond.cuh
+++ b/timemachine/cpp/src/kernels/k_flat_bottom_bond.cuh
@@ -8,7 +8,13 @@ template <typename RealType>
 RealType __device__ __forceinline__ compute_flat_bottom_energy(RealType k, RealType r, RealType rmin, RealType rmax) {
     RealType r_gt_rmax = static_cast<RealType>(r > rmax);
     RealType r_lt_rmin = static_cast<RealType>(r < rmin);
-    return (k / 4) * (r_lt_rmin * (pow(r - rmin, 4)) + r_gt_rmax * (pow(r - rmax, 4)));
+    RealType d_rmin_2 = (r - rmin) * (r - rmin);
+    RealType d_rmin_4 = d_rmin_2 * d_rmin_2;
+
+    RealType d_rmax_2 = (r - rmax) * (r - rmax);
+    RealType d_rmax_4 = d_rmax_2 * d_rmax_2;
+
+    return (k / static_cast<RealType>(4.0)) * ((r_lt_rmin * d_rmin_4) + (r_gt_rmax * d_rmax_4));
 }
 
 template <typename RealType>
@@ -98,6 +104,7 @@ void __global__ k_flat_bottom_bond(
     // compute common subexpressions involving distance, displacements
     RealType dx[3];
     RealType r2 = 0;
+#pragma unroll
     for (int d = 0; d < 3; d++) {
         double delta = coords[src_idx * 3 + d] - coords[dst_idx * 3 + d];
         delta -= box[d * 3 + d] * nearbyint(delta / box[d * 3 + d]);
@@ -115,39 +122,48 @@ void __global__ k_flat_bottom_bond(
         // Always set the energy buffer value to ensure buffer is initialized
         u[b_idx] = FLOAT_TO_FIXED_ENERGY<RealType>(u_real);
     }
-
-    if (du_dp) {
-        // compute parameter derivatives
-        RealType du_dk_real = (r_gt_rmax * (pow(r - rmax, 4) / 4)) + (r_lt_rmin * (pow(r - rmin, 4) / 4));
-        RealType du_drmin_real = r_lt_rmin * (-k * pow(r - rmin, 3));
-        RealType du_drmax_real = r_gt_rmax * (-k * pow(r - rmax, 3));
-
-        // cast float -> fixed
-        auto du_dk = FLOAT_TO_FIXED_BONDED<RealType>(du_dk_real);
-        auto du_drmin = FLOAT_TO_FIXED_BONDED<RealType>(du_drmin_real);
-        auto du_drmax = FLOAT_TO_FIXED_BONDED<RealType>(du_drmax_real);
-
-        // increment du_dp array
-        atomicAdd(du_dp + k_idx, du_dk);
-        atomicAdd(du_dp + rmin_idx, du_drmin);
-        atomicAdd(du_dp + rmax_idx, du_drmax);
-    }
-
-    if (du_dx) {
-        RealType du_dr = k * ((r_gt_rmax * pow(r - rmax, 3)) + (r_lt_rmin * pow(r - rmin, 3)));
-
-        for (int d = 0; d < 3; d++) {
-            // compute du/dcoords
-            RealType du_dsrc_real = du_dr * dx[d] / r;
-            RealType du_ddst_real = -du_dsrc_real;
+    if (du_dp || du_dx) {
+        RealType d_r_min = r - rmin;
+        RealType d_r_max = r - rmax;
+        RealType d_rmin_3 = d_r_min * d_r_min * d_r_min;
+        RealType d_rmax_3 = d_r_max * d_r_max * d_r_max;
+        if (du_dp) {
+            // compute parameter derivatives
+            RealType du_dk_real =
+                (r_gt_rmax * ((d_rmax_3 * d_r_max) / static_cast<RealType>(4)) +
+                 (r_lt_rmin * ((d_rmin_3 * d_r_min) / static_cast<RealType>(4))));
+            RealType du_drmin_real = r_lt_rmin * (-k * d_rmin_3);
+            RealType du_drmax_real = r_gt_rmax * (-k * d_rmax_3);
 
             // cast float -> fixed
-            auto du_dsrc = FLOAT_TO_FIXED_BONDED<RealType>(du_dsrc_real);
-            auto du_ddst = FLOAT_TO_FIXED_BONDED<RealType>(du_ddst_real);
+            unsigned long long du_dk = FLOAT_TO_FIXED_BONDED<RealType>(du_dk_real);
+            unsigned long long du_drmin = FLOAT_TO_FIXED_BONDED<RealType>(du_drmin_real);
+            unsigned long long du_drmax = FLOAT_TO_FIXED_BONDED<RealType>(du_drmax_real);
 
-            // increment du_dx array
-            atomicAdd(du_dx + src_idx * 3 + d, du_dsrc);
-            atomicAdd(du_dx + dst_idx * 3 + d, du_ddst);
+            // increment du_dp array
+            atomicAdd(du_dp + k_idx, du_dk);
+            atomicAdd(du_dp + rmin_idx, du_drmin);
+            atomicAdd(du_dp + rmax_idx, du_drmax);
+        }
+
+        if (du_dx) {
+            RealType du_dr = k * ((r_gt_rmax * d_rmax_3) + (r_lt_rmin * d_rmin_3));
+
+            RealType inv_r = 1 / r;
+#pragma unroll
+            for (int d = 0; d < 3; d++) {
+                // compute du/dcoords
+                RealType du_dsrc_real = du_dr * dx[d] * inv_r;
+                RealType du_ddst_real = -du_dsrc_real;
+
+                // cast float -> fixed
+                unsigned long long du_dsrc = FLOAT_TO_FIXED_BONDED<RealType>(du_dsrc_real);
+                unsigned long long du_ddst = FLOAT_TO_FIXED_BONDED<RealType>(du_ddst_real);
+
+                // increment du_dx array
+                atomicAdd(du_dx + src_idx * 3 + d, du_dsrc);
+                atomicAdd(du_dx + dst_idx * 3 + d, du_ddst);
+            }
         }
     }
 }