--- a/timemachine/cpp/src/local_md_potentials.cu
+++ b/timemachine/cpp/src/local_md_potentials.cu
@@ -55,7 +55,7 @@ LocalMDPotentials::LocalMDPotentials(
         default_bonds[i * 2 + 1] = i + 1;
     }
     std::vector<double> default_params(N_ * 3);
-    free_restraint_ = std::shared_ptr<FlatBottomBond<double>>(new FlatBottomBond<double>(default_bonds));
+    free_restraint_ = std::shared_ptr<FlatBottomBond<float>>(new FlatBottomBond<float>(default_bonds));
     // Construct a bound potential with 0 params
     bound_free_restraint_ = std::shared_ptr<BoundPotential>(new BoundPotential(free_restraint_, default_params));
 
@@ -71,8 +71,8 @@ LocalMDPotentials::LocalMDPotentials(
     all_potentials_.push_back(bound_free_restraint_);
     all_potentials_.push_back(ixn_group_);
     if (!freeze_reference) {
-        frozen_restraint_ = std::shared_ptr<LogFlatBottomBond<double>>(
-            new LogFlatBottomBond<double>(default_bonds, 1 / (temperature * BOLTZ)));
+        frozen_restraint_ = std::shared_ptr<LogFlatBottomBond<float>>(
+            new LogFlatBottomBond<float>(default_bonds, 1 / (temperature * BOLTZ)));
         bound_frozen_restraint_ =
             std::shared_ptr<BoundPotential>(new BoundPotential(frozen_restraint_, default_params));
         all_potentials_.push_back(bound_frozen_restraint_);