--- a/timemachine/cpp/src/local_md_potentials.hpp
+++ b/timemachine/cpp/src/local_md_potentials.hpp
@@ -57,11 +57,11 @@ private:
     std::shared_ptr<BoundPotential> ixn_group_;
     std::shared_ptr<BoundPotential> nonbonded_bp_;
     // Restraint for the free particles to the reference particle
-    std::shared_ptr<FlatBottomBond<double>> free_restraint_;
+    std::shared_ptr<FlatBottomBond<float>> free_restraint_;
     std::shared_ptr<BoundPotential> bound_free_restraint_;
 
     // Restraint for the frozen particles to the reference particle
-    std::shared_ptr<LogFlatBottomBond<double>> frozen_restraint_;
+    std::shared_ptr<LogFlatBottomBond<float>> frozen_restraint_;
     std::shared_ptr<BoundPotential> bound_frozen_restraint_;
 
     DeviceBuffer<int> d_restraint_pairs_;