@badisa 2024-04-29
> Nice improvements!
> 
> Observation: In the posted timings, `k_log_flat_bottom_bond` is more expensive than `k_flat_bottom_bond`. Don't know if that's mostly because `k_log_flat_bottom_bond` is being invoked on a larger number of pairs (`n_pairs = n_atoms - n_local_atoms - 1` vs. `n_pairs = n_local_atoms - 1`), or if the arithmetic cost per pair is higher, or if the cost of the competing atomic adds is higher (in each case, I believe each coord of the central atom needs to receive `n_pairs` atomicAdds?)... In a later pass, it may be good to re-profile at a range of `n_atoms`, `n_local_atoms`...

Taking a look at the log flat bottom, I noticed there were two exp calls when there could be one (about 200-500ns faster). Fixed in https://github.com/proteneer/timemachine/pull/1297/commits/13b751c0a4626730d13947ad16d66d3f486e8257 and also avoided 3 divisions in favor of 1 diversions and 3 multiplies. The second didn't really make a difference, but to be consistent with the flat bottom version